Comments - Defect Number: 3369; Defect RecordID: 3365; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: Margin Visibility issue; Defect Description: Working with Anna --- when she looks P001218  Digital Marketing
Financial Summary  Hidden margins
Budget Mgmt --- Can see all information
Burn Report  Hidden margins
Budget Comparison  can see all information

This is a BIG security issue ---- all reports should have margin hidden

-EllenB


Hi Nick,

We identified that the issue is because of the flags to be set depending on the role were not set accordingly. We resolved it and we will deploy this to UAT today.

Thanks