Comments - Defect Number: 3383; Defect RecordID: 3379; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: P001360 project fix for EIC null value.; Defect Description: P001360
Sharath,

Can you look into why I am getting the following error when attempting to see Project Detail view in PM? I just built this project out a couple weeks ago, and it was working fine at that time.

Thanks!

-NickC

Hi Nick,

The project P001360, does not have any EIC assigned to it and the project has a Margin Exception. So, while checking for Logged in user is EIC, we are getting an error (as a null value is stored for EIC). We have fixed this issue in Local. Please let us know if we can move this to Production as we do not have similar project in UAT.

Thanks
Sharath Chilakala

Nick callahan

Approved.
On Nov 13, 2017, at 5:18 AM, Sharath. Chilakala <sharath.chilakala@ggktech.com> wrote: